# Wedding/Wedding.API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Wedding.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Activity": {
      "get": {
        "tags": [
          "Activity"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "filterOn",
            "type": "string"
          },
          {
            "in": "query",
            "name": "filterQuery",
            "type": "string"
          },
          {
            "in": "query",
            "name": "sortBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "isAscending",
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "pageNumber",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Activity"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateActivityLogDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Activity/{id}": {
      "get": {
        "tags": [
          "Activity"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Activity"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateActivityLogDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Activity"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateActivityLogDTO": {
      "type": "object",
      "properties": {
        "logId": {
          "format": "uuid",
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "action": {
          "type": "string"
        },
        "entity": {
          "type": "string"
        },
        "timeStamp": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "ResponseDTO": {
      "type": "object",
      "properties": {
        "result": { },
        "isSuccess": {
          "type": "boolean"
        },
        "statusCode": {
          "format": "int32",
          "type": "integer"
        },
        "message": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateActivityLogDTO": {
      "type": "object",
      "properties": {
        "logId": {
          "format": "uuid",
          "type": "string"
        },
        "action": {
          "type": "string"
        },
        "entity": {
          "type": "string"
        },
        "timeStamp": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Please enter your token with this format: \"Bearer YOUR_TOKEN\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}